remove deals rpc, client, and cli code
