+++ ffs/rpc/rpc.proto
@@ -3,8 +3,6 @@ package ffs.rpc;
 
 option go_package = "github.com/textileio/powergate/ffs/rpc";
 
-import "deals/rpc/rpc.proto";
-
 // model messages
 // ------------------------------------
 
@@ -151,6 +149,56 @@ enum Direction {
    string addr = 2;
    Direction direction = 3;
  }
+
+ message ListDealRecordsConfig {
+   repeated string from_addrs = 1;
+   repeated string data_cids = 2;
+   bool include_pending = 3;
+   bool include_final = 4;
+   bool ascending = 5;
+}
+
+message StorageDealInfo {
+	string proposal_cid = 1;
+	uint64 state_id = 2;
+	string state_name = 3;
+	string miner = 4;
+
+	string piece_cid = 5;
+	uint64 size = 6;
+
+	uint64 price_per_epoch = 7;
+	uint64 start_epoch = 8;
+	uint64 duration = 9;
+
+	uint64 deal_id = 10;
+	int64 activation_epoch = 11;
+	string msg = 12;
+}
+
+message StorageDealRecord {
+   string root_cid = 1;
+   string addr = 2;
+   int64 time = 3;
+   bool pending = 4;
+   StorageDealInfo deal_info = 5;
+}
+
+message RetrievalDealInfo {
+	string root_cid = 1;
+	uint64 size = 2;
+	uint64 min_price = 3;
+	uint64 payment_interval = 4;
+	uint64 payment_interval_increase = 5;
+	string miner = 6;
+	string miner_peer_id = 7;
+}
+
+message RetrievalDealRecord {
+    string addr = 1;
+    int64 time = 2;
+    RetrievalDealInfo deal_info = 3;
+}
 
 // request/reply messages
 // -------------------------------------
@@ -349,19 +397,19 @@ message ListPayChannelsRequest {
  }
 
  message ListStorageDealRecordsRequest {
-   deals.rpc.ListDealRecordsConfig config = 1;
+   ListDealRecordsConfig config = 1;
 }
 
 message ListStorageDealRecordsResponse {
-   repeated deals.rpc.StorageDealRecord records = 1;
+   repeated StorageDealRecord records = 1;
 }
 
 message ListRetrievalDealRecordsRequest {
-   deals.rpc.ListDealRecordsConfig config = 1;
+   ListDealRecordsConfig config = 1;
 }
 
 message ListRetrievalDealRecordsResponse {
-   repeated deals.rpc.RetrievalDealRecord records = 1;
+   repeated RetrievalDealRecord records = 1;
 }
 
 message ShowAllRequest {
